pyproject.toml
[project]
name = "rag_kitcore"
version = "0.1.0"
description = "Local, privacy-first RAG framework with hybrid retrieval and Ollama"
requires-python = ">=3.12"
authors = [{ name = "Simone" }]

dependencies = [
    # Core RAG stack
    "sentence-transformers",
    "qdrant-client",
    "rank-bm25",
    "langchain",
    "langchain-community",
    "langchain-core",
    "langchain-ollama",
    "langchain-openai",

    # Embeddings & reranker
    "numpy",
    "scikit-learn",

    # Web UI
    "streamlit",

    # Config & utilities
    "pydantic",
    "python-dotenv",
    "pyyaml",
    "requests",
]

[tool.ruff]
line-length = 100
